backend/paddle_service/requirements.txt
Flask==2.3.3
paddleocr>=2.7.3
paddlepaddle-gpu>=2.6.0
PyMuPDF>=1.24.0
opencv-python==4.9.0.80
Pillow>=10.2.0
numpy>=1.21.0
gunicorn
